+++ filament/backend/src/metal/MetalDriver.mm
@@ -209,42 +209,39 @@ void MetalDriver::createRenderTargetR(Handle<HwRenderTarget> rth,
         uint8_t samples, backend::MRT color,
         TargetBufferInfo depth, TargetBufferInfo stencil) {
 
-    auto getColorTexture = [&]() -> id<MTLTexture> {
-        if (color[0].handle) {
-            auto colorTexture = handle_cast<MetalTexture>(mHandleMap, color[0].handle);
-            ASSERT_PRECONDITION(colorTexture->texture,
-                    "Color texture passed to render target has no texture allocation");
-            colorTexture->updateLodRange(color[0].level);
-            return colorTexture->texture;
-        } else if (any(targetBufferFlags & TargetBufferFlags::COLOR)) {
-            ASSERT_POSTCONDITION(false, "The COLOR flag was specified, but no color texture provided.");
+    MetalRenderTarget::Attachment colorAttachments[4] = {{ 0 }};
+    for (size_t i = 0; i < 4; i++) {
+        const auto& buffer = color[i];
+        if (!buffer.handle) {
+            ASSERT_POSTCONDITION(none(targetBufferFlags & getMRTColorFlag(i)),
+                    "The COLOR%d flag was specified, but no color texture provided.", i);
+            continue;
         }
-        return nil;
-    };
 
-    auto getDepthTexture = [&]() -> id<MTLTexture> {
-        if (depth.handle) {
-            auto depthTexture = handle_cast<MetalTexture>(mHandleMap, depth.handle);
-            ASSERT_PRECONDITION(depthTexture->texture,
-                    "Depth texture passed to render target has no texture allocation.");
-            depthTexture->updateLodRange(depth.level);
-            return depthTexture->texture;
-        } else if (any(targetBufferFlags & TargetBufferFlags::DEPTH)) {
-            ASSERT_POSTCONDITION(false, "The DEPTH flag was specified, but no depth texture provided.");
-        }
-        return nil;
-    };
+        auto colorTexture = handle_cast<MetalTexture>(mHandleMap, buffer.handle);
+        ASSERT_PRECONDITION(colorTexture->texture,
+                "Color texture passed to render target has no texture allocation");
+        colorTexture->updateLodRange(buffer.level);
+        colorAttachments[i].texture = colorTexture->texture;
+        colorAttachments[i].level = color[0].level;
+        colorAttachments[i].layer = color[0].layer;
+    }
 
-    MetalRenderTarget::TargetInfo colorInfo;
-    colorInfo.level = color[0].level;
-    colorInfo.layer = color[0].layer;
-
-    MetalRenderTarget::TargetInfo depthInfo;
-    depthInfo.level = depth.level;
-    depthInfo.layer = depth.layer;
+    MetalRenderTarget::Attachment depthAttachment = { 0 };
+    if (depth.handle) {
+        auto depthTexture = handle_cast<MetalTexture>(mHandleMap, depth.handle);
+        ASSERT_PRECONDITION(depthTexture->texture,
+                "Depth texture passed to render target has no texture allocation.");
+        depthTexture->updateLodRange(depth.level);
+        depthAttachment.texture = depthTexture->texture;
+        depthAttachment.level = depth.level;
+        depthAttachment.layer = depth.layer;
+    }
+    ASSERT_POSTCONDITION(!depth.handle || any(targetBufferFlags & TargetBufferFlags::DEPTH),
+            "The DEPTH flag was specified, but no depth texture provided.");
 
     construct_handle<MetalRenderTarget>(mHandleMap, rth, mContext, width, height, samples,
-            getColorTexture(), colorInfo, getDepthTexture(), depthInfo);
+            colorAttachments, depthAttachment);
 
     ASSERT_POSTCONDITION(
             !stencil.handle &&
@@ -699,36 +696,7 @@ void MetalDriver::beginRenderPass(Handle<HwRenderTarget> rth,
     mContext->currentRenderPassFlags = params.flags;
 
     MTLRenderPassDescriptor* descriptor = [MTLRenderPassDescriptor renderPassDescriptor];
-
-    const auto discardFlags = params.flags.discardEnd;
-    const bool discardColor = any(discardFlags & TargetBufferFlags::COLOR);
-    const bool discardDepth = any(discardFlags & TargetBufferFlags::DEPTH);
-
-    // Color
-
-    const auto& colorAttachment = descriptor.colorAttachments[0];
-    colorAttachment.texture = renderTarget->getColor();
-    colorAttachment.resolveTexture = discardColor ? nil : renderTarget->getColorResolve();
-    colorAttachment.level = renderTarget->getColorInfo().level;
-    colorAttachment.slice = renderTarget->getColorInfo().layer;
-    mContext->currentSurfacePixelFormat = colorAttachment.texture.pixelFormat;
-
-    colorAttachment.loadAction = renderTarget->getLoadAction(params, TargetBufferFlags::COLOR);
-    colorAttachment.storeAction = renderTarget->getStoreAction(params, TargetBufferFlags::COLOR);
-    colorAttachment.clearColor = MTLClearColorMake(
-            params.clearColor.r, params.clearColor.g, params.clearColor.b, params.clearColor.a);
-
-    // Depth
-
-    const auto& depthAttachment = descriptor.depthAttachment;
-    depthAttachment.texture = renderTarget->getDepth();
-    depthAttachment.resolveTexture = discardDepth ? nil : renderTarget->getDepthResolve();
-    depthAttachment.loadAction = renderTarget->getLoadAction(params, TargetBufferFlags::DEPTH);
-    depthAttachment.storeAction = renderTarget->getStoreAction(params, TargetBufferFlags::DEPTH);
-    depthAttachment.clearDepth = params.clearDepth;
-    depthAttachment.level = renderTarget->getDepthInfo().level;
-    depthAttachment.slice = renderTarget->getDepthInfo().layer;
-    mContext->currentDepthPixelFormat = descriptor.depthAttachment.texture.pixelFormat;
+    renderTarget->setUpRenderPassAttachments(descriptor, params);
 
     mContext->currentRenderPassEncoder =
             [getPendingCommandBuffer(mContext) renderCommandEncoderWithDescriptor:descriptor];
@@ -850,8 +818,10 @@ void MetalDriver::readPixels(Handle<HwRenderTarget> src, uint32_t x, uint32_t y,
                         "readPixels must be called outside of a render pass.");
 
     auto srcTarget = handle_cast<MetalRenderTarget>(mHandleMap, src);
-    id<MTLTexture> srcTexture = srcTarget->getColor();
-    size_t miplevel = srcTarget->getColorInfo().level;
+    // We always readPixels from the COLOR0 attachment.
+    MetalRenderTarget::Attachment color = srcTarget->getColorAttachment(0);
+    id<MTLTexture> srcTexture = color.texture;
+    size_t miplevel = color.level;
 
     auto chooseMetalPixelFormat = [] (PixelDataFormat format, PixelDataType type) {
         // TODO: Add support for UINT and INT
@@ -938,8 +908,12 @@ void MetalDriver::blit(TargetBufferFlags buffers,
                         dstRect.left >= 0 && dstRect.bottom >= 0,
             "Source and destination rects must be positive.");
 
-    id<MTLTexture> srcTexture = srcTarget->getBlitColorSource();
-    id<MTLTexture> dstTexture = dstTarget->getColor();
+    // We always blit to/from the COLOR0 attachment.
+    MetalRenderTarget::Attachment srcColorAttachment = srcTarget->getColorAttachment(0);
+    MetalRenderTarget::Attachment dstColorAttachment = dstTarget->getColorAttachment(0);
+
+    id<MTLTexture> srcTexture = srcColorAttachment.texture;
+    id<MTLTexture> dstTexture = dstColorAttachment.texture;
 
     ASSERT_PRECONDITION(srcTexture != nil && dstTexture != nil,
             "Source texture and destination texture must not be nil");
@@ -956,8 +930,8 @@ void MetalDriver::blit(TargetBufferFlags buffers,
             dstTexture.height - (NSUInteger) dstRect.bottom - dstRect.height,
             dstRect.width, dstRect.height);
 
-    const uint8_t srcLevel = srcTarget->getColorInfo().level;
-    const uint8_t dstLevel = dstTarget->getColorInfo().level;
+    const uint8_t srcLevel = srcColorAttachment.level;
+    const uint8_t dstLevel = dstColorAttachment.level;
 
     auto isBlitableTextureType = [](MTLTextureType t) {
         return t == MTLTextureType2D || t == MTLTextureType2DMultisample;
@@ -979,8 +953,10 @@ void MetalDriver::blit(TargetBufferFlags buffers,
     }
 
     if (any(buffers & TargetBufferFlags::DEPTH)) {
-        args.source.depth = srcTarget->getBlitDepthSource();
-        args.destination.depth = dstTarget->getDepth();
+        MetalRenderTarget::Attachment srcDepthAttachment = srcTarget->getDepthAttachment();
+        MetalRenderTarget::Attachment dstDepthAttachment = dstTarget->getDepthAttachment();
+        args.source.depth = srcDepthAttachment.texture;
+        args.destination.depth = dstDepthAttachment.texture;
     }
 
     mContext->blitter->blit(getPendingCommandBuffer(mContext), args);
@@ -994,12 +970,30 @@ void MetalDriver::draw(backend::PipelineState ps, Handle<HwRenderPrimitive> rph)
     const auto& rs = ps.rasterState;
 
     // Pipeline state
+    MTLPixelFormat colorPixelFormat[4] = { MTLPixelFormatInvalid };
+    for (size_t i = 0; i < 4; i++) {
+        const auto& attachment = mContext->currentRenderTarget->getColorAttachment(i);
+        if (!attachment) {
+            continue;
+        }
+        colorPixelFormat[i] = attachment.texture.pixelFormat;
+    }
+    MTLPixelFormat depthPixelFormat = MTLPixelFormatInvalid;
+    const auto& depthAttachment = mContext->currentRenderTarget->getDepthAttachment();
+    if (depthAttachment) {
+        depthPixelFormat = depthAttachment.texture.pixelFormat;
+    }
     metal::PipelineState pipelineState {
         .vertexFunction = program->vertexFunction,
         .fragmentFunction = program->fragmentFunction,
         .vertexDescription = primitive->vertexDescription,
-        .colorAttachmentPixelFormat = mContext->currentSurfacePixelFormat,
-        .depthAttachmentPixelFormat = mContext->currentDepthPixelFormat,
+        .colorAttachmentPixelFormat = {
+            colorPixelFormat[0],
+            colorPixelFormat[1],
+            colorPixelFormat[2],
+            colorPixelFormat[3]
+        },
+        .depthAttachmentPixelFormat = depthPixelFormat,
         .sampleCount = mContext->currentRenderTarget->getSamples(),
         .blendState = BlendState {
             .blendingEnabled = rs.hasBlending(),